moved chore and chore list to packages
